Switch to a richer annotation format.
The richer format contains the search score in addition to the final
score, and also contains some other metadata.

And I think there was also a bug in the old annotator: positions were
first filtered, then alternating scores were assigned to the filtered
positions. This can lead to a parity error if an intermediate position
is filtered (though this is rare). This bug is now fixed.