Switch to DTZ-optimal play after 1st repetition.
Change Syzygy root filtering to apply the 'hasRepeated' condition if
there have been any repetitions since the last irreversible move. This
is in contrast to the current behavior of only setting that condition if
the *current* position has occurred before.

This should cause the engine to switch to DTZ-optimal play if it is in a
position that is won according to the EGTB, but where a repetition has
occurred in play (showing that it is not making progress). This should
prevent situations where a won position results in a draw due to 3-fold
repetition.