Improve MY_ASSERT macro a bit (#71)
Now avoids evaluating its condition twice in debug mode, while still ensuring that control-flow analysis tools (in particular, clang-tidy-21) recognize that the condition is guaranteed to hold after the assertion (in debug mode).

Also, using `std::abort` to trigger failure in `consteval` contexts feels nicer and more readable than relying on a `nullptr` dereference.

Apparently with VS2026 (or maybe Windows 11?) it's no longer necessary to call `_set_error_mode(_OUT_TO_MSGBOX)` to get the debugger to trigger on assertion failures. Without that call the assertion failure is printed to the terminal and a message box still appears with a 'retry' button to trigger the debugger.

Removing the repeated calls to `_set_error_mode` and the duplicate condition evaluations slightly improves performance under debug mode.

Confirmed that we're still clang-tidy-clean if upgrading from clang-tidy-19 to 21 (#66).